backend/requirements.txt

flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-bcrypt
psycopg2-binary
python-dotenv
marshmallow-sqlalchemy
flask-jwt-extended
flask-limiter
redis
celery
redis
requests
python-dateutil
marshmallow
flask-caching
flask-mail
python-multipart
pyotp
qrcode
pillow
pandas
numpy
scikit-learn
prophet
statsmodels
scipy
